+++ 1D/TASEP.cpp
@@ -34,8 +34,8 @@ int main(int argc, char **argv) {
     I.show();
     for (size_t t = 0, u = 0;; ++t, ++u) {
         for (size_t j = 0; j < u; ++j)
-            for (unsigned i = 0; i < T.size(); ++i) T.step();
-        for (unsigned i = 0; i < T.size(); ++i) I.put({i, t % size_t(I.h())}, C[T[i]]);
+            for (size_t i = 0; i < T.size(); ++i) T.step();
+        for (size_t i = 0; i < T.size(); ++i) I.put({i, t % size_t(I.h())}, C[T[i]]);
         if (t == size_t(I.h()) - 1) I.pause();
     }
 }
