+++ 2D/Perco-Schramm.cpp
@@ -78,7 +78,7 @@ private:
     size_t follow(size_t base, size_t dir) {
         static const vector<int> fola = {1, int(w), int(w) - 1, -1, -int(w) - 1, -int(w)};
         static const vector<int> folb = {1, int(w) + 1, int(w), -1, -int(w), -int(w) + 1};
-        return size_t((((base / w) % 2) != 0 ? folb : fola)[dir] + int(base));
+        return size_t((((base / w) % 2) != 0 ? folb : fola)[dir] + int64_t(base));
     }
 
     void seg(Path *p, size_t base, size_t dir, size_t rot) {
